requirement.txt
asgiref==3.8.1
Django==4.2.18
django-admin-interface==0.29.4
django-admin-rangefilter==0.13.2
django-colorfield==0.11.0
django-cors-headers==4.6.0
djangorestframework==3.15.2
pillow==11.1.0
python-slugify==8.0.4
pytz==2024.2
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.12.2
tzdata==2024.2
